# docs/swagger.yaml
definitions:
...
  controllers.DeleteResponse:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
  controllers.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  models.Comment:
    properties:
      created_at:
        type: string
      id:
        type: integer
      message:
        type: string
      photo:
        $ref: '#/definitions/models.Photo'
      photoID:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      userID:
        type: integer
    type: object
  models.Photo:
    properties:
      caption:
        type: string
      created_at:
        type: string
      id:
        type: integer
      photo_url:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      userID:
        type: integer
    type: object
  models.SocialMedia:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      social_media_url:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      userID:
        type: integer
    type: object
  models.User:
    properties:
      age:
        type: integer
      ban_reason:
        type: string
      banned_at:
        type: string
      comments:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      last_login_at:
        type: string
      last_seen_at:
        type: string
      photos:
        items:
          $ref: '#/definitions/models.Photo'
        type: array
      role:
        type: string
      social_medias:
        items:
          $ref: '#/definitions/models.SocialMedia'
        type: array
      status:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/v1/photos:
    get:
      consumes:
      - application/json
      description: Get all existing photos
      produces:
      - application/json
      responses:
        "200":
          description: Get all photos success
          schema:
            items:
              $ref: '#/definitions/models.Photo'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Database error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "503":
          description: Database is not ready
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all photos
      tags:
      - photo
    post:
      consumes:
      - application/json
      description: Create photo to post in finalproject
      parameters:
      - description: Photo request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Photo'
      produces:
      - application/json
      responses:
        "201":
          description: Create photo success
          schema:
            $ref: '#/definitions/models.Photo'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "503":
          description: Database is not ready
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create photo
      tags:
      - photo
  /api/v1/photos/{photoId}:
    delete:
      consumes:
      - application/json
      description: Delete photo identified by given ID
      parameters:
      - description: ID of the photo
        in: path
        name: photoId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Delete photo success
          schema:
            $ref: '#/definitions/controllers.DeleteResponse'
        "400":
          description: Invalid photo id
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Photo not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "503":
          description: Database is not ready
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete photo
      tags:
      - photo
    get:
      consumes:
      - application/json
      description: Get photo by ID
      parameters:
      - description: ID of the photo
        in: path
        name: photoId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Get photo success
          schema:
            $ref: '#/definitions/models.Photo'
        "400":
          description: Invalid photo id
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Photo not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Database error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "503":
          description: Database is not ready
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get photo
      tags:
      - photo
    put:
      consumes:
      - application/json
      description: Update photo identified by given ID
      parameters:
      - description: ID of the photo
        in: path
        name: photoId
        required: true
        type: integer
      - description: Photo update request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Photo'
      produces:
      - application/json
      responses:
        "200":
          description: Update photo success
          schema:
            $ref: '#/definitions/models.Photo'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Photo not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "503":
          description: Database is not ready
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update photo
      tags:
      - photo
...
  /photos/create:
    post:
      consumes:
      - application/json
      description: Create photo to post in finalproject
      parameters:
      - description: Photo request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Photo'
      produces:
      - application/json
      responses:
        "201":
          description: Create photo success
          schema:
            $ref: '#/definitions/models.Photo'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "503":
          description: Database is not ready
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create photo
      tags:
      - photo
  /photos/delete/{photoId}:
    delete:
      consumes:
      - application/json
      description: Delete photo identified by given ID
      parameters:
      - description: ID of the photo
        in: path
        name: photoId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Delete photo success
          schema:
            $ref: '#/definitions/controllers.DeleteResponse'
        "400":
          description: Invalid photo id
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Photo not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "503":
          description: Database is not ready
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete photo
      tags:
      - photo
  /photos/get/{photoId}:
    get:
      consumes:
      - application/json
      description: Get photo by ID
      parameters:
      - description: ID of the photo
        in: path
        name: photoId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Get photo success
          schema:
            $ref: '#/definitions/models.Photo'
        "400":
          description: Invalid photo id
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Photo not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Database error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "503":
          description: Database is not ready
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get photo
      tags:
      - photo
  /photos/getall:
    get:
      consumes:
      - application/json
      description: Get all existing photos
      produces:
      - application/json
      responses:
        "200":
          description: Get all photos success
          schema:
            items:
              $ref: '#/definitions/models.Photo'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Database error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "503":
          description: Database is not ready
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all photos
      tags:
      - photo
  /photos/update/{photoId}:
    put:
      consumes:
      - application/json
      description: Update photo identified by given ID
      parameters:
      - description: ID of the photo
        in: path
        name: photoId
        required: true
        type: integer
      - description: Photo update request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Photo'
      produces:
      - application/json
      responses:
        "200":
          description: Update photo success
          schema:
            $ref: '#/definitions/models.Photo'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Photo not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "503":
          description: Database is not ready
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update photo
      tags:
      - photo
...
swagger: "2.0"